Rename schemas/users -> user
